refactor: 컬럼 추가 버튼 로직 분리
